+++ packages/shared/src/components/squads/cards/feed/PublicSquadEligibilityCard.tsx
